refactor(footer): Footer 컴포넌트 구조 리팩터링 및 하위 컴포넌트 분리
- FooterTop : 푸터의 상단 영역으로, 각종 정보/서비스 링크 그룹과 앱 다운로드 아이콘을 정렬된 UI로 표시하는 컴포넌트
- FooterBottom : 푸터의 하단 영역으로, 회사 정보, 약관, 연락처, 저작권 및 SNS 링크 아이콘들을 정렬하는 컴포넌트
