+++ src/layout/Footer/FooterBottom.module.css
@@ -1,0 +1,15 @@
+.FooterLink {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    background-color: #ced4da;
+}
+
+.Icon {
+    width: 20px;
+    height: 20px;
+    fill: '#2c2e33';
+}
