+++ src/layout/Footer/FooterTop.jsx
@@ -1,0 +1,85 @@
+import {
+    Box,
+    Group,
+    Text,
+    Anchor,
+    Stack,
+    Container,
+    rem
+} from '@mantine/core';
+
+const FooterTop = () => {
+    return <Box>
+        <Container size='xl'>
+            <Group align='flex-start' justify='space-between' wrap='nowrap' mb={30}>
+                {[
+                    {
+                        title: '인프런',
+                        items: [
+                            ['인프런 소개', 'https://www.inflearn.com/inflearn'],
+                            ['인프런 피드', 'https://www.inflearn.com/pages'],
+                            ['수강평 모아보기', 'https://www.inflearn.com/community/reviews'],
+                            ['블로그', 'https://www.inflearn.com/blogs'],
+                        ],
+                    },
+                    {
+                        title: '신청하기',
+                        items: [
+                            ['지식공유참여', 'https://www.inflearn.com/open-knowledge'],
+                            ['멘토링 소개', 'https://www.inflearn.com/mentors-intro'],
+                            ['인프런 비즈니스', 'https://www.inflearn.com/intro-group'],
+                            ['인프런 제휴', 'https://www.inflearn.com/partners'],
+                        ],
+                    },
+                    {
+                        title: '코드 등록',
+                        items: [
+                            ['수강코드 등록', 'https://www.inflearn.com/course-code'],
+                            ['포인트코드 등록', 'https://www.inflearn.com/point-code'],
+                        ],
+                    },
+                    {
+                        title: '고객센터',
+                        items: [
+                            ['공지사항', 'https://www.inflearn.com/notices'],
+                            ['자주 묻는 질문', 'https://www.inflearn.com/faq'],
+                            ['저작권 신고센터', 'https://www.inflearn.com/copyright'],
+                            ['수료증 확인', 'https://www.inflearn.com/certificate'],
+                            ['강의 · 기능요청', 'https://www.inflearn.com/request-to-inflearn'],
+                        ],
+                    },
+                    {
+                        title: '인프랩',
+                        items: [
+                            ['인프랩 실Log', 'https://doc.clickup.com/d/3gfz7-5843/log/3gfz7-75205'],
+                            ['With us', 'https://www.inflearn.com/pages/withus'],
+                            ['인프랩 스토리', 'https://story.inflab.com/'],
+                            ['인프랩 테크', 'https://tech.inflab.com/'],
+                            ['IT 인재 채용 서비스', 'https://business.rallit.com/'],
+                        ],
+                    },
+                ].map((group, idx) => (
+                    <Stack gap={6} key={idx} style={{ minWidth: rem(120) }}>
+                        <Text fw={600} fz={14} c='white' mb={4}>{group.title}</Text>
+                        {group.items.map(([label, href], i) => (
+                            <Anchor key={i} href={href} c='#ced4da' fz={14} display='block'>
+                                {label}
+                            </Anchor>
+                        ))}
+                    </Stack>
+                ))}
+
+                <Stack align='center' gap={8} style={{ minWidth: rem(90), flexShrink: 0 }}>
+                    <svg width='60' height='60' viewBox='0 0 80 80' fill='none'>
+                        <path d='M4.364 4.364h15.272v15.273H4.364zm2.181 2.182v10.909h10.91V6.545zm2.182 2.181h6.546v6.546H8.727zm43.637-4.363h15.272v15.273H52.364zm2.181 2.182v10.909h10.91V6.545zm2.182 2.181h6.546v6.546h-6.546zM4.364 52.364h15.272v15.273H4.364zm2.181 2.182v10.909h10.91v-10.91zm2.182 2.181h6.546v6.546H8.727zM26.182 4.364h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm10.91 0h2.18v2.182h-2.18zm-24 2.182h2.18v2.181h-2.18zm10.908 0h2.182v2.181h-2.182zm8.727 0h2.182v2.181h-2.182zm4.364 0H48v2.181h-2.182zm2.182 0h2.182v2.181H48zM26.182 8.727h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.182zm2.181 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H37.09zm6.545 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H48zM24 10.91h2.182v2.182H24zm2.182 0h2.182v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm6.546 0h2.181v2.182h-2.181zm4.363 0h2.182v2.182h-2.182zm2.182 0H48v2.182h-2.182zm-24 2.182H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.546 0h2.181v2.182h-2.181zm6.545 0H48v2.182h-2.182zM34.91 15.273h2.182v2.182H34.91zm2.182 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm4.364 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm-26.182 2.182H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H34.91zm4.364 0h2.181v2.182h-2.181zm4.363 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H48zm-24 2.182h2.182v2.181H24zm2.182 0h2.182v2.181h-2.182zm2.182 0h2.181v2.181h-2.182zm2.181 0h2.182v2.181h-2.182zm6.546 0h2.182v2.181H37.09zm4.363 0h2.182v2.181h-2.182zm6.546 0h2.182v2.181H48zM4.364 21.818h2.181V24H4.364zm4.363 0h2.182V24H8.727zm4.364 0h2.182V24H13.09zm4.363 0h2.182V24h-2.182zm6.546 0h2.182V24H24zm6.545 0h2.182V24h-2.182zm8.728 0h2.181V24h-2.181zm2.181 0h2.182V24h-2.182zm2.182 0h2.182V24h-2.182zm2.182 0H48V24h-2.182zm2.182 0h2.182V24H48zm8.727 0h2.182V24h-2.182zm6.546 0h2.181V24h-2.181zM4.363 24h2.182v2.182H4.364zm10.91 0h2.181v2.182h-2.181zm4.363 0h2.182v2.182h-2.182zM24 24h2.182v2.182H24zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm10.909 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm6.545 0h2.182v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zM6.545 26.182h2.182v2.182H6.545zm6.546 0h2.182v2.182H13.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0H24v2.182h-2.182zm2.182 0h2.182v2.182H24zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm10.909 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm4.363 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H61.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm-61.09 2.182h2.181v2.182H4.364zm2.181 0h2.182v2.182H6.545zm4.364 0h2.182v2.182h-2.182zm8.727 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm8.727 0h2.182v2.182h-2.182zm6.546 0h2.181v2.182h-2.181zM6.545 30.546h2.182v2.181H6.545zm2.182 0h2.182v2.181H8.727zm2.182 0h2.182v2.181h-2.182zm4.364 0h2.181v2.181h-2.181zm2.181 0h2.182v2.181h-2.182zm2.182 0h2.182v2.181h-2.182zm8.728 0h2.181v2.181h-2.182zm2.181 0h2.182v2.181h-2.182zm4.364 0h2.182v2.181h-2.182zm2.182 0h2.182v2.181H37.09zm2.182 0h2.181v2.181h-2.181zm2.181 0h2.182v2.181h-2.182zm4.364 0H48v2.181h-2.182zm2.182 0h2.182v2.181H48zm2.182 0h2.182v2.181h-2.182zm2.182 0h2.181v2.181h-2.181zm6.545 0h2.182v2.181h-2.182zm4.364 0h2.181v2.181h-2.181zm2.181 0h2.182v2.181h-2.182zm-61.09 2.181h2.181v2.182H4.364zm4.363 0h2.182v2.182H8.727zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm6.545 0H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182H37.09zm4.363 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H48zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm6.545 0h2.182v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zM4.364 34.91h2.181v2.182H4.364zm2.181 0h2.182v2.182H6.545zm6.546 0h2.182v2.182H13.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H24zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.546 0h2.181v2.182h-2.181zm13.09 0h2.182v2.182h-2.181zm6.546 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zM10.91 37.091h2.182v2.182H10.91zm8.727 0h2.182v2.182h-2.182zm2.182 0H24v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H48zm10.909 0h2.182v2.182H58.91zm4.364 0h2.181v2.182h-2.181zM6.545 39.273h2.182v2.182H6.545zm2.182 0h2.182v2.182H8.727zm4.364 0h2.182v2.182H13.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm4.364 0H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm13.09 0h2.182v2.182h-2.181zm2.182 0h2.182v2.182h-2.182zm4.364 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm6.545 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zM8.727 41.455h2.182v2.182H8.727zm6.546 0h2.181v2.182h-2.181zm6.545 0H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.182zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.546 0h2.181v2.182h-2.181zm8.727 0h2.182v2.182H48zm4.364 0h2.181v2.182h-2.181zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182H61.09zm4.363 0h2.182v2.182h-2.182zm-61.09 2.182h2.181v2.181H4.364zm8.727 0h2.182v2.181H13.09zm4.363 0h2.182v2.181h-2.182zm2.182 0h2.182v2.181h-2.182zm2.182 0H24v2.181h-2.182zm2.182 0h2.182v2.181H24zm2.182 0h2.182v2.181h-2.182zm4.363 0h2.182v2.181h-2.182zm4.364 0h2.182v2.181h-2.182zm6.545 0h2.182v2.181h-2.182zm10.91 0h2.181v2.181h-2.181zm10.909 0h2.181v2.181h-2.181zm2.181 0h2.182v2.181h-2.182zm-58.909 2.18h2.182V48H6.545zm2.182 0h2.182V48H8.727zm4.364 0h2.182V48H13.09zm2.182 0h2.181V48h-2.181zm4.363 0h2.182V48h-2.182zm4.364 0h2.182V48H24zm8.727 0h2.182V48h-2.182zm2.182 0h2.182V48h-2.182zm2.182 0h2.182V48H37.09zm4.363 0h2.182V48h-2.182zm2.182 0h2.182V48h-2.182zm6.546 0h2.182V48h-2.182zm2.182 0h2.181V48h-2.181zm4.363 0h2.182V48h-2.182zm2.182 0h2.182V48h-2.182zm4.364 0h2.181V48h-2.181zM4.363 48h2.182v2.182H4.364zm13.091 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zM48 48h2.182v2.182H48zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm-34.909 2.182H24v2.182h-2.182zm4.364 0h2.182v2.182h-2.182zm8.727 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm8.727 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H61.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm-39.272 2.182h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm8.727 0h2.182v2.182h-2.182zm4.364 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm4.364 0h2.181v2.182h-2.181zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zM24 54.546h2.182v2.181H24zm2.182 0h2.182v2.181h-2.182zm2.182 0h2.181v2.181h-2.182zm10.909 0h2.181v2.181h-2.181zm2.181 0h2.182v2.181h-2.182zm2.182 0h2.182v2.181h-2.182zm4.364 0h2.182v2.181H48zm8.727 0h2.182v2.181h-2.182zm2.182 0h2.182v2.181h-2.182zm-37.09 2.181h2.18v2.182h-2.18zm2.18 0h2.183v2.182H24zm13.092 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H48zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm8.727 0h2.182v2.182h-2.182zM26.182 58.91h2.182v2.182h-2.182zm2.182 0h2.181v2.182h-2.182zm2.181 0h2.182v2.182h-2.182zm8.728 0h2.181v2.182h-2.181zm6.545 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H61.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm-43.636 2.182H24v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm8.728 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm4.364 0H48v2.182h-2.182zm8.727 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm6.545 0h2.182v2.182h-2.182zm-39.272 2.182h2.182v2.182h-2.182zm4.363 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182h-2.182zm4.364 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm4.364 0H48v2.182h-2.182zm2.182 0h2.182v2.182H48zm4.364 0h2.181v2.182h-2.181zm10.909 0h2.181v2.182h-2.181zm-41.455 2.182H24v2.182h-2.182zm2.182 0h2.182v2.182H24zm10.909 0h2.182v2.182h-2.182zm2.182 0h2.182v2.182H37.09zm2.182 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182zm6.546 0h2.182v2.182H48zm15.273 0h2.181v2.182h-2.181zm2.181 0h2.182v2.182h-2.182z' fill='#CED4DA'></path>
+                    </svg>
+                    <Text fz={12} fw={600} c='white'>
+                        Get the app
+                    </Text>
+                </Stack>
+            </Group>
+        </Container>
+    </Box>
+}
+
+export default FooterTop;
